# v3: Optimize IsFromLocal() performance