--- a/ctx_test.go
+++ b/ctx_test.go
@@ -6352,3 +6352,61 @@ func Benchmark_Ctx_IsProxyTrusted(b *testing.B) {
 		})
 	})
 }
+
+func Benchmark_Ctx_IsFromLocalhost(b *testing.B) {
+	// Scenario without localhost check
+	b.Run("Non_Localhost", func(b *testing.B) {
+		app := New()
+		c := app.AcquireCtx(&fasthttp.RequestCtx{})
+		c.Request().SetRequestURI("http://google.com:8080/test")
+		b.ReportAllocs()
+		b.ResetTimer()
+		for n := 0; n < b.N; n++ {
+			c.IsFromLocal()
+		}
+		app.ReleaseCtx(c)
+	})
+
+	// Scenario without localhost check in parallel
+	b.Run("Non_Localhost_Parallel", func(b *testing.B) {
+		app := New()
+		b.ReportAllocs()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			c := app.AcquireCtx(&fasthttp.RequestCtx{})
+			c.Request().SetRequestURI("http://google.com:8080/test")
+			for pb.Next() {
+				c.IsFromLocal()
+			}
+			app.ReleaseCtx(c)
+		})
+	})
+
+	// Scenario with localhost check
+	b.Run("Localhost", func(b *testing.B) {
+		app := New()
+		c := app.AcquireCtx(&fasthttp.RequestCtx{})
+		c.Request().SetRequestURI("http://localhost:8080/test")
+		b.ReportAllocs()
+		b.ResetTimer()
+		for n := 0; n < b.N; n++ {
+			c.IsFromLocal()
+		}
+		app.ReleaseCtx(c)
+	})
+
+	// Scenario with localhost check in parallel
+	b.Run("Localhost_Parallel", func(b *testing.B) {
+		app := New()
+		b.ReportAllocs()
+		b.ResetTimer()
+		b.RunParallel(func(pb *testing.PB) {
+			c := app.AcquireCtx(&fasthttp.RequestCtx{})
+			c.Request().SetRequestURI("http://localhost:8080/test")
+			for pb.Next() {
+				c.IsFromLocal()
+			}
+			app.ReleaseCtx(c)
+		})
+	})
+}