# Description

- Add Benchmarks for `ctx.IsFromLocal()`
- Optimize performance of `ctx.IsFromLocal()` by using `net.IP` functions to detect if IP is from loopback address.
- Reduce memory allocations by using `net.IP` directly instead of strings.

## Benchmark Results:

| **Benchmark**                  | **Before ns/op** | **After ns/op** | **Speedup**     | **Before Allocations** | **After Allocations** | **Improvement**        |
|--------------------------------|-----------------|-----------------|-----------------|-----------------------|-----------------------|-------------------------|
| **Non_Localhost**              | 26.45           | 7.51            | ~3.52× faster    | 1 allocs/op            | 0 allocs/op            | Allocations Eliminated  |
| **Non_Localhost_Parallel**     | 6.824           | 1.30            | ~5.25× faster    | 1 allocs/op            | 0 allocs/op            | Allocations Eliminated  |
| **Localhost**                  | 26.48           | 7.52            | ~3.52× faster    | 1 allocs/op            | 0 allocs/op            | Allocations Eliminated  |
| **Localhost_Parallel**         | 6.974           | 1.31            | ~5.33× faster    | 1 allocs/op            | 0 allocs/op            | Allocations Eliminated  |


## Changes introduced

- [x] Benchmarks: Describe any performance benchmarks and improvements related to the changes.
- [x] API Longevity: Discuss the steps taken to ensure that the new or updated APIs are consistent and not prone to breaking changes.

## Type of change

- [x] Performance improvement (non-breaking change which improves efficiency)