@gaby 2024-09-23
> # ⚠️ **Performance Alert** ⚠️
> Possible performance regression was detected for benchmark. Benchmark result of this commit is worse than the previous benchmark result exceeding threshold `1.50`.
> 
> Benchmark suite	Current: [7b13bfe](https://github.com/gofiber/fiber/commit/7b13bfe5cc26fe00e768961a96f1a31f9051d2a2)	Previous: [fbc24e8](https://github.com/gofiber/fiber/commit/fbc24e83d63653237decfe86a6684bc5a417ff28)	Ratio
> `BenchmarkAppendMsgitem - MB/s`	`3209.82` MB/s	`1847.42` MB/s	`1.74`
> `BenchmarkAppendMsgitem - MB/s`	`2878.85` MB/s	`1847.42` MB/s	`1.56`
> This comment was automatically generated by [workflow](https://github.com/gofiber/fiber/actions?query=workflow%3ABenchmark) using [github-action-benchmark](https://github.com/marketplace/actions/continuous-benchmark).

This is related to this https://github.com/benchmark-action/github-action-benchmark/issues/264